Move bubble event into seperate file
